+++ packages/rrweb-snapshot/package.json
@@ -54,6 +54,7 @@
   },
   "homepage": "https://github.com/rrweb-io/rrweb/tree/master/packages/rrweb-snapshot#readme",
   "devDependencies": {
+    "@rrweb/types": "^2.0.0-alpha.17",
     "@rrweb/utils": "^2.0.0-alpha.17",
     "@types/jsdom": "^20.0.0",
     "@types/node": "^18.15.11",
