+++ packages/rrweb-snapshot/test/utils.test.ts
@@ -2,14 +2,14 @@
  * @vitest-environment jsdom
  */
 import { describe, it, test, expect } from 'vitest';
-import { NodeType, serializedNode } from '../src/types';
 import {
   escapeImportStatement,
   extractFileExtension,
   fixSafariColons,
   isNodeMetaEqual,
 } from '../src/utils';
-import type { serializedNodeWithId } from 'rrweb-snapshot';
+import { NodeType } from '@rrweb/types';
+import type { serializedNode, serializedNodeWithId } from '@rrweb/types';
 
 describe('utils', () => {
   describe('isNodeMetaEqual()', () => {
