+++ packages/rrweb-snapshot/tsconfig.json
@@ -8,6 +8,9 @@
   },
   "references": [
     {
+      "path": "../types"
+    },
+    {
       "path": "../utils"
     }
   ]
